(DM-2466) define LSSTSW as relative to the the ./etc/ or ./bin/ dir
This change is intended to make the path selection behavior consistent with
existing user expectations that the stack installation will be relative to the
path where the lsstsw repo is cloned.  Note that we are both removing the
requirement that the user set LSSTSW in their env and shadowing (ignoring) any
value they may have exported.