split miniconda & conda pkg install into separate steps
Prior to this change, if `deploy` successfully installed miniconda but a
failure occurred while installing conda packages, the install could not
be repaired by rerunning `deploy`. This was due to the miniconda
installer refusing to run if the install directory already exists.  This
is avoided by protecting the miniconda and conda package installs with
separate lockfiles.