point daf_butlerUtils and obs_base to new link.
We need daf_butlerUtils to exist in repos.yaml so that old tags still build.
It's safe to point it to obs_base, because our build system uses the name in
repos.yaml as the name of the clone/eups product, so old versions (pre-rename)
will work with the old name.