remove original bNNNN when publishing a DISTRIBTAG
At present, when publishing a symbolic "DISTRIBTAG", E.g, w_2017_14,
both it at the bNNNN tag it was copied from are [publicly] published,
resulting in a 2x faster tag growth rate.  As the bNNNN tag is not
intended to be consumed by an end-user, it is safe to remove.

As EUPS distrib downloads all available tags and labels installed
products with them, it is desirable to limit the number of tags for
both performance and to reduce the amount of meta-data the end-user is
subject too.